+++ clients/js/networks.ts
@@ -99,8 +99,8 @@ const MAINNET = {
     key: undefined,
   },
   moonbeam: {
-    rpc: undefined,
-    key: undefined,
+    rpc: "https://rpc.api.moonbeam.network",
+    key: get_env_var("ETH_KEY"),
   },
   neon: {
     rpc: undefined,
@@ -110,6 +110,18 @@ const MAINNET = {
     rpc: "https://phoenix-lcd.terra.dev",
     chain_id: "phoenix-1",
     key: get_env_var("TERRA_MNEMONIC"),
+  },
+  arbitrum: {
+    rpc: "https://arb1.arbitrum.io/rpc",
+    key: get_env_var("ETH_KEY"),
+  },
+  optimism: {
+    rpc: "https://mainnet.optimism.io",
+    key: get_env_var("ETH_KEY"),
+  },
+  gnosis: {
+    rpc: "https://rpc.gnosischain.com/",
+    key: get_env_var("ETH_KEY"),
   },
   ropsten: {
     rpc: `https://rpc.ankr.com/eth_ropsten`,
@@ -212,6 +224,18 @@ const TESTNET = {
     chain_id: "pisco-1",
     key: get_env_var("TERRA_MNEMONIC"),
   },
+  arbitrum: {
+    rpc: "https://rinkeby.arbitrum.io/rpc",
+    key: get_env_var("ETH_KEY_TESTNET"),
+  },
+  optimism: {
+    rpc: "https://goerli.optimism.io",
+    key: get_env_var("ETH_KEY_TESTNET"),
+  },
+  gnosis: {
+    rpc: "https://sokol.poa.network/",
+    key: get_env_var("ETH_KEY_TESTNET"),
+  },
   ropsten: {
     rpc: `https://rpc.ankr.com/eth_ropsten`,
     key: get_env_var("ETH_KEY_TESTNET"),
@@ -312,6 +336,18 @@ const DEVNET = {
     rpc: "http://localhost:1318",
     chain_id: "phoenix-1",
     key: "notice oak worry limit wrap speak medal online prefer cluster roof addict wrist behave treat actual wasp year salad speed social layer crew genius",
+  },
+  arbitrum: {
+    rpc: undefined,
+    key: undefined,
+  },
+  optimism: {
+    rpc: undefined,
+    key: undefined,
+  },
+  gnosis: {
+    rpc: undefined,
+    key: undefined,
   },
   ropsten: {
     rpc: undefined,