+++ node/cmd/guardiand/node.go
@@ -26,6 +26,7 @@ import (
 	"github.com/certusone/wormhole/node/pkg/common"
 	"github.com/certusone/wormhole/node/pkg/devnet"
 	"github.com/certusone/wormhole/node/pkg/ethereum"
+	"github.com/certusone/wormhole/node/pkg/governor"
 	"github.com/certusone/wormhole/node/pkg/p2p"
 	"github.com/certusone/wormhole/node/pkg/processor"
 	gossipv1 "github.com/certusone/wormhole/node/pkg/proto/gossip/v1"
@@ -127,6 +128,10 @@ var (
 	solanaWsRPC *string
 	solanaRPC   *string
 
+	pythnetContract *string
+	pythnetWsRPC    *string
+	pythnetRPC      *string
+
 	logLevel *string
 
 	unsafeDevMode   *bool
@@ -154,6 +159,8 @@ var (
 	bigTableTableName          *string
 	bigTableTopicName          *string
 	bigTableKeyPath            *string
+
+	chainGovernorEnabled *bool
 )
 
 func init() {
@@ -235,6 +242,10 @@ func init() {
 	solanaWsRPC = NodeCmd.Flags().String("solanaWS", "", "Solana Websocket URL (required")
 	solanaRPC = NodeCmd.Flags().String("solanaRPC", "", "Solana RPC URL (required")
 
+	pythnetContract = NodeCmd.Flags().String("pythnetContract", "", "Address of the PythNet program (required)")
+	pythnetWsRPC = NodeCmd.Flags().String("pythnetWS", "", "PythNet Websocket URL (required")
+	pythnetRPC = NodeCmd.Flags().String("pythnetRPC", "", "PythNet RPC URL (required")
+
 	logLevel = NodeCmd.Flags().String("logLevel", "info", "Logging level (debug, info, warn, error, dpanic, panic, fatal)")
 
 	unsafeDevMode = NodeCmd.Flags().Bool("unsafeDevMode", false, "Launch node in unsafe, deterministic devnet mode")
@@ -266,6 +277,8 @@ func init() {
 	bigTableTableName = NodeCmd.Flags().String("bigTableTableName", "", "BigTable table name to store events in")
 	bigTableTopicName = NodeCmd.Flags().String("bigTableTopicName", "", "GCP topic name to publish to")
 	bigTableKeyPath = NodeCmd.Flags().String("bigTableKeyPath", "", "Path to json Service Account key")
+
+	chainGovernorEnabled = NodeCmd.Flags().Bool("chainGovernorEnabled", false, "Run the chain governor")
 }
 
 var (
@@ -352,6 +365,9 @@ func runNode(cmd *cobra.Command, args []string) {
 	readiness.RegisterComponent(common.ReadinessEthSyncing)
 	if *solanaWsRPC != "" {
 		readiness.RegisterComponent(common.ReadinessSolanaSyncing)
+	}
+	if *pythnetWsRPC != "" {
+		readiness.RegisterComponent(common.ReadinessPythNetSyncing)
 	}
 	if *terraWS != "" {
 		readiness.RegisterComponent(common.ReadinessTerraSyncing)
@@ -616,6 +632,16 @@ func runNode(cmd *cobra.Command, args []string) {
 			if *algorandAppID == 0 {
 				logger.Fatal("Please specify --algorandAppID")
 			}
+
+			if *pythnetContract == "" {
+				logger.Fatal("Please specify --pythnetContract")
+			}
+			if *pythnetWsRPC == "" {
+				logger.Fatal("Please specify --pythnetWsUrl")
+			}
+			if *pythnetRPC == "" {
+				logger.Fatal("Please specify --pythnetUrl")
+			}
 		}
 
 	}
@@ -678,6 +704,13 @@ func runNode(cmd *cobra.Command, args []string) {
 	solAddress, err := solana_types.PublicKeyFromBase58(*solanaContract)
 	if err != nil {
 		logger.Fatal("invalid Solana contract address", zap.Error(err))
+	}
+	var pythnetAddress solana_types.PublicKey
+	if *testnetMode {
+		pythnetAddress, err = solana_types.PublicKeyFromBase58(*pythnetContract)
+		if err != nil {
+			logger.Fatal("invalid PythNet contract address", zap.Error(err))
+		}
 	}
 
 	// In devnet mode, we generate a deterministic guardian key and write it to disk.
@@ -773,6 +806,7 @@ func runNode(cmd *cobra.Command, args []string) {
 		chainObsvReqC[vaa.ChainIDNeon] = make(chan *gossipv1.ObservationRequest)
 		chainObsvReqC[vaa.ChainIDEthereumRopsten] = make(chan *gossipv1.ObservationRequest)
 		chainObsvReqC[vaa.ChainIDInjective] = make(chan *gossipv1.ObservationRequest)
+		chainObsvReqC[vaa.ChainIDPythNet] = make(chan *gossipv1.ObservationRequest)
 	}
 
 	// Multiplex observation requests to the appropriate chain
@@ -858,14 +892,28 @@ func runNode(cmd *cobra.Command, args []string) {
 	// provides methods for reporting progress toward message attestation, and channels for receiving attestation lifecyclye events.
 	attestationEvents := reporter.EventListener(logger)
 
-	publicrpcService, publicrpcServer, err := publicrpcServiceRunnable(logger, *publicRPC, db, gst)
+	var gov *governor.ChainGovernor
+	if *chainGovernorEnabled {
+		logger.Info("chain governor is enabled")
+		env := governor.MainNetMode
+		if *testnetMode {
+			env = governor.TestNetMode
+		} else if *unsafeDevMode {
+			env = governor.DevNetMode
+		}
+		gov = governor.NewChainGovernor(logger, db, env)
+	} else {
+		logger.Info("chain governor is disabled")
+	}
+
+	publicrpcService, publicrpcServer, err := publicrpcServiceRunnable(logger, *publicRPC, db, gst, gov)
 
 	if err != nil {
 		log.Fatal("failed to create publicrpc service socket", zap.Error(err))
 	}
 
 	// local admin service socket
-	adminService, err := adminServiceRunnable(logger, *adminSocketPath, injectC, signedInC, obsvReqSendC, db, gst)
+	adminService, err := adminServiceRunnable(logger, *adminSocketPath, injectC, signedInC, obsvReqSendC, db, gst, gov)
 	if err != nil {
 		logger.Fatal("failed to create admin service socket", zap.Error(err))
 	}
@@ -879,7 +927,7 @@ func runNode(cmd *cobra.Command, args []string) {
 	// Run supervisor.
 	supervisor.New(rootCtx, logger, func(ctx context.Context) error {
 		if err := supervisor.Run(ctx, "p2p", p2p.Run(
-			obsvC, obsvReqC, obsvReqSendC, sendC, signedInC, priv, gk, gst, *p2pPort, *p2pNetworkID, *p2pBootstrap, *nodeName, *disableHeartbeatVerify, rootCtxCancel)); err != nil {
+			obsvC, obsvReqC, obsvReqSendC, sendC, signedInC, priv, gk, gst, *p2pPort, *p2pNetworkID, *p2pBootstrap, *nodeName, *disableHeartbeatVerify, rootCtxCancel, gov)); err != nil {
 			return err
 		}
 
@@ -988,13 +1036,32 @@ func runNode(cmd *cobra.Command, args []string) {
 
 		if *solanaWsRPC != "" {
 			if err := supervisor.Run(ctx, "solwatch-confirmed",
-				solana.NewSolanaWatcher(*solanaWsRPC, *solanaRPC, solAddress, lockC, nil, rpc.CommitmentConfirmed).Run); err != nil {
+				solana.NewSolanaWatcher(*solanaWsRPC, *solanaRPC, solAddress, lockC, nil, rpc.CommitmentConfirmed, common.ReadinessSolanaSyncing, vaa.ChainIDSolana).Run); err != nil {
 				return err
 			}
 
 			if err := supervisor.Run(ctx, "solwatch-finalized",
-				solana.NewSolanaWatcher(*solanaWsRPC, *solanaRPC, solAddress, lockC, chainObsvReqC[vaa.ChainIDSolana], rpc.CommitmentFinalized).Run); err != nil {
+				solana.NewSolanaWatcher(*solanaWsRPC, *solanaRPC, solAddress, lockC, chainObsvReqC[vaa.ChainIDSolana], rpc.CommitmentFinalized, common.ReadinessSolanaSyncing, vaa.ChainIDSolana).Run); err != nil {
 				return err
+			}
+		}
+
+		if *pythnetWsRPC != "" {
+			if err := supervisor.Run(ctx, "pythwatch-confirmed",
+				solana.NewSolanaWatcher(*pythnetWsRPC, *pythnetRPC, pythnetAddress, lockC, nil, rpc.CommitmentConfirmed, common.ReadinessPythNetSyncing, vaa.ChainIDPythNet).Run); err != nil {
+				return err
+			}
+
+			if err := supervisor.Run(ctx, "pythwatch-finalized",
+				solana.NewSolanaWatcher(*pythnetWsRPC, *pythnetRPC, pythnetAddress, lockC, chainObsvReqC[vaa.ChainIDPythNet], rpc.CommitmentFinalized, common.ReadinessPythNetSyncing, vaa.ChainIDPythNet).Run); err != nil {
+				return err
+			}
+		}
+
+		if gov != nil {
+			err := gov.Run(ctx)
+			if err != nil {
+				log.Fatal("failed to create chain governor", zap.Error(err))
 			}
 		}
 
@@ -1013,6 +1080,7 @@ func runNode(cmd *cobra.Command, args []string) {
 			*ethRPC,
 			attestationEvents,
 			notifier,
+			gov,
 		)
 		if err := supervisor.Run(ctx, "processor", p.Run); err != nil {
 			return err